renamed and added some classes
