+++ src/main/java/com/zipcodewilmington/scientificcalculator/Console.java
@@ -12,6 +12,8 @@ import java.lang.Math;
  */
 public class Console {
 
+    static double storage;
+
     public static void print(String output, Object... args) {
         System.out.printf(output, args);
     }
@@ -94,6 +96,16 @@ public class Console {
     public static double clearScreen() {
         System.out.flush();
         return 0.0;
+    }
+
+    public static void save(double Number) {
+        storage = Number;
+        System.out.println(storage);
+    }
+
+    public static double recall() {
+        return storage;
+
     }
