+++ src/main/java/com/zipcodewilmington/scientificcalculator/TrigFunctions.java
@@ -1,6 +1,6 @@
 package com.zipcodewilmington.scientificcalculator;
 
-public class ScientificFunctions {
+public class TrigFunctions {
 
     public static Double sine(Double number, String mode) {
         if (mode.equals("degrees")) {
@@ -44,36 +44,5 @@ public class ScientificFunctions {
         return Math.atan(number);
     }
 
-    public static Double factorial(Double userInput) {
-        if (userInput == 0) {
-            return 1.0;
-        }
-        else {
-            Double result = 1.0;
-            for (int i = 1; i < userInput; i++) {
-                result = result * (i + 1);
-            }
-            return result;
-        }
-    }
 
-    public static Double logarithm(Double userInput) {
-        Double log = Math.log10(userInput);
-        return log;
-    }
-
-    public static Double inverseLogarithm (Double userInput){
-        Double inverseLog = Math.pow(10, userInput);
-        return inverseLog;
-    }
-
-    public static Double naturalLogarithm(Double userInput) {
-        Double natLog = Math.log(userInput);
-        return natLog;
-    }
-
-    public static Double inverseNaturalLogarithm(Double userInput) {
-        Double inverseNatLog = Math.exp(userInput);
-        return inverseNatLog;
-    }
 }
